Conflicts:
	teamProject/src/main/java/ebook/EBookCategoryServ.java
	teamProject/src/main/java/ebook/EBookDAO.java
	teamProject/src/main/webapp/ebook/eBookCategory.jsp

+++ teamProject/src/main/java/member/MemberFindEmailServ.java
@@ -63,11 +63,11 @@ public class MemberFindEmailServ implements Controller
 			switch (rIndex) {
 			case 0:
 				//a-z
-				temp.append((char) ((int) (rnd.nextInt(26)) + 97));
+				temp.append((char) (rnd.nextInt(26)) + 97);
 				 break;
             case 1:
                 // A-Z
-                temp.append((char) ((int) (rnd.nextInt(26)) + 65));
+                temp.append((char) (rnd.nextInt(26)) + 65);
                 break;
             case 2:
                 // 0-9
